Apply commit to update ~.xcodeproj file to unstash the changes
There were some issues with PictureSlippingTableViewCells project, so it was jist deleted and created the new one with the same name. All the files in previous commit idnetical according to git except PictureSlippingTableViewCells.xcodeproj which was marked as 'modified'. Apply this commit to update the new ictureSlippingTableViewCells.xcodeproj. It needs for unstashing the changes made in the PictureSlippingTableViewCells