Remove class factor method in User, and make some other refactor changes
Remove class factor method 'userWithName:image:' in User class, as it may need to write additional unit tests for the method which is unneeded because we already have the 'initWithName:image' method.

Redefine 'usersArray' and 'userNamesArray' variables as properties in UsersListViewController class because it may cause data leaks when setting the variables from two different instances of the class.

And also make some other minor refactor changes and perfomance improvements, such as changing [NSMutableArray new] method to [NSMutableArray arrayWithCapacity:] in 'makeUsersArrayWithCount:' method