Replace "broken" Reminders project with working one
It seems Reminders project to be broken after creating the PictureSlippingTableViewCells project.
The thing is when we have created the PictureSlippingTableViewCells project, some files from the Reminders project was copied to the first one. But Xcode instead of copying the files moved them to the PictureSlippingTableViewCells project and created references in the Reminders project to them.
During the work on the PictureSlippingTableViewCells project these files was modified permanently and can't be used to reproduce previous state of the Reminders project.
This commit deletes broken Reminders project and replaces it with the working one, which was get by cloning repository and reseting first 4 commits from HEAD in it.