Integrate biometry authentication into lock screen
Integrate biometry authentication with Face ID or Touch ID into the RILockScreenViewController using LocalAuthentication framework.

Also create 'hideBiometryButton' method in RINumberPad and 'hiddenAppearance' property in RINumberPadButton. The last one hides the appearance of the button, but not the view of the button itself. It's needed because when using Apple-provided 'hidden' property it causes removing the button from the 'numberPadStackView' and, as a result, incorrectly aligned buttons in the number pad.