Store reminders' image attachments as files in /Documents directory
Store reminders' image attachments as files in /Documents directory. Images are stored as .png files.
Path to image attachment is ~/Documents/ImageAttachments/<reminder-ID>/<image-ID>.png, where reminder-ID and image-ID are UUID-based strings; in current realization these IDs are used only for navigating in the file system.
Also in current realization these attachments are cleaned when RITasksListViewController deallocates, because between launches of the app there's no place where we can store URLs to the images, therefore it may lead to memory leaks on SSD.