Implement custom URL scheme for Reminders app, RICreateReminderDelegate
Implement custom URL scheme for Reminders app. The structure of the string is as follows: "reminders:REMINDERS_BODY?imagesAr[0]=BASE64_DATA_0&imagesAr[1]=BASE64+DATA_1&...", where REMINDERS_BODY is the text of the reminder; BASE64_DATA_0, BASE64_DATA_1, and so on are Base64 encoded strings with image datas.

It's important to follow these conventions when naming query arguments of the array:

- parameter name must be 'imagesAr';
- next you should append non-negative number in brackets. For example 'imagesAr[2]', which means the 3rd item of the array;
- the last index of the images array must be less than total number of the 'imagesAr' arguments per 1. For example, if you pass 4 items of the array to URL, your query parameters must look like:
"?imagesAr[0]=...&imagesAr[1]=...&imagesAr[2]=...&imagesAr[3]=...".

Violations of this rules will lead to crash when parsing the URL inside the Reminders app.

The following combinations are not allowed, but not limited to: 'imagesAr', 'imagesAr[', 'imagesAr[]', 'imagesAr[-7]', 'imagesAr[dsfgfd]', etc.

Also the app, that will call this URL scheme, have to include in Info.plist file 'reminders' value for LSApplicationQueriesSchemes key. The type of the value must be Array.

Also create RICreateReminderDelegate and apply it with RICreateReminderViewController to be able to setup the view controller in SceneDelegate.