Create RIEntryPasscodeViewController and RISecureManager
Create RIEntryPasscodeViewController with 3 options of entry: setting passcode, entering passcode to confirm access, and changing old passcode to some new one.

Also create RISecureManager, which encapsulates storing and managing access to the passcode logic. All the operations with the passcode sych as validating, changing, setting and reseting must be done using this abstraction.
This class is a singletone object, which can be accessed through +shared method.