Integrate Core Data into the app, minor refactor changes
Integrate Core Data into the app. Now there're 2 models, which are very similar: RIReminder and RIReminderRaw. RIReminderRaw is actually almost like RIReminder in the previous versions of the code before this commit. RIReminder here is a model purposed to be stored in Core Data (it's subclass of NSManagedObject).

Create RICoreDataStack class which will work with initializing and preparing Core Data stack to work. Include it in RIAppDelegate and when you will need to access the RICoreDataStack instance inject it from RIAppDelegate instance to your class.

Also make some minor refactor changes due to code style.