Normalize return types of +instance methods, refactor change
Normalize return types of +instance methods in all view controllers. Return exactly that type which view controller's type it is. Before this commit there were some view controllers which returned UINavigationController from them which might be confusing sometimes. Now if you want your controller to be wrapped into UINavigationController do this explicitly in the place you want it.

Make some refactor changes, in particular rename all +shared methods to +sharedInstance. Then refactor them to readonly class properties 'sharedInstance' except RIPasscodeManager. Here leave +sharedInstance method because of convenience conforming to protocol RIPasscodeManagerProtocol: we can specify 'instancetype' as return type of the method and not to bound on which type this method returns (in Objective-C we don't have such a great mechanism as associated type of protocol in Swift and when declaring class property in Objective-C protocol we will have to explictly specify which type this property must return).